+++ Frameworks/OWF/Processors.subproj/OWProcessor.m
@@ -1,4 +1,4 @@
-// Copyright 1997-2005, 2007, 2010 Omni Development, Inc.  All rights reserved.
+// Copyright 1997-2005, 2007, 2010-2011 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -268,12 +268,12 @@ unlockAndReturn:
     return [pipeline firstBytesDate];
 }
 
-- (unsigned int)bytesProcessed;
+- (NSUInteger)bytesProcessed;
 {
     return [pipeline bytesProcessed];
 }
 
-- (unsigned int)totalBytes;
+- (NSUInteger)totalBytes;
 {
     return [pipeline totalBytes];
 }
